Fix compatibility issues with Yii `22.0` and PHP `8.5`.

+++ src/ColumnSchema.php
@@ -107,7 +107,7 @@ class ColumnSchema extends BaseObject
             case 'boolean':
                 return (bool) $value;
             case 'double':
-                return (double) $value;
+                return (float) $value;
         }
 
         return $value;
